Added verification of the status code and the response body to each test. Deleted extra tests. Divided the class ScooterAPI into two parts: CourierAPI and ScooterAPI.
