+++ src/test/java/GetListOfOrdersTest.java
@@ -1,15 +1,16 @@
-import scooter.staticmethodsandvariables.ScooterAPI;
+import scooter.staticmethodsandvariables.OrderAPI;
 import org.junit.Test;
 
 import static org.apache.http.HttpStatus.SC_OK;
 import static org.hamcrest.CoreMatchers.notNullValue;
 
 public class GetListOfOrdersTest {
-    ScooterAPI scooterAPI = new ScooterAPI();
+    OrderAPI orderAPI = new OrderAPI();
+
     @Test
-    public void doesGetListOfOrdersReturnListOfOrdersAndCode200(){
-        scooterAPI.getListOfOrders().then().statusCode(SC_OK)
+    public void doesGetListOfOrdersReturnListOfOrdersAndCode200() {
+        orderAPI.getListOfOrders().then().statusCode(SC_OK)
                 .and()
-                .body("orders",notNullValue());
+                .body("orders", notNullValue());
     }
 }
